Move event listener logic from grid to AppCanvas
